+++ Source/SelectClassTraits/ChangeLog.md
@@ -1,4 +1,10 @@
 ﻿ChangeLog of Select Class Traits, a Phoenix Point Mod by Lotrick87
+# Version 2.0, 30-01-2021
+-Complete rework of Modnix / Config
+-Choose any skill(not only default passive traits)
+-option to make classes static(skills are generated in order by numbers, not randomly)
+-multiple builds per class support
+-opttion to Leave empty slots
 
 # Version 1.0.1, 25-01-2021
 -Updated Readme File